1.0-night.rev4
Removing DOMContentLoaded event listener.

Still stuck on figuring out how to identify if clients died or not,
would be best if on unload I could get a message so I can remove it from
aArrClientIds.

Also I'm not able to send a message to the clientMessageListener for
serverRequest_toUpdatedServer_unregisterCompleted.